requirements.txt
pylint
pytest
pytest-cov
python-dotenv
kagglehub
torchvision
torch
matplotlib
seaborn
tqdm
scikit-learn
Pillow